requirements.txt
Flask>=2.0
Flask-SQLAlchemy
Flask-Bcrypt
pymysql
